Initial commit of multiple file inputs for concatenation.
This branch seeks to expand the cli to allow multiple video files to be specified. After adding OSD and SRT back, it will concatenate the files
with FFMPEG. At this point in the commit it has been barely tested but seems to work.

If concatenating, the intermediate video files will be deleted.